requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
Django==5.1.4
django-resized==1.0.3
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
Pygments==2.18.0
pyparsing==3.2.1
python-telegram-bot==21.10
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tiktok_downloader==0.3.5
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
yt-dlp==2024.12.23
